[restructure] move vcard related utils to lib/utils/vcard
